Split analyzer definition & formatters into separate files of the same partial class
